Revert "feat: return error code if failed"

+++ examples/queue_multiple_transactions_and_poll_all_results/queue_multiple_transactions_and_poll_all_results_slave/queue_multiple_transactions_and_poll_all_results_slave.ino
@@ -58,7 +58,7 @@ void loop()
         Serial.println("all queued transactions completed. start verifying received data from slave");
 
         // get received bytes for all transactions
-        const std::vector<int64_t> received_bytes = slave.numBytesReceivedAll();
+        const std::vector<size_t> received_bytes = slave.numBytesReceivedAll();
 
         // verify and dump difference with received data
         // NOTE: we need only 1st results (received_bytes[0])
